[FIX] l10n_mx_edi_landing: customs number unique by company
In Mexican localization, landed costs have an extra field,
l10n_mx_edi_customs_number, that has a sql unique constraint.

Before this commit, the customs number was unique in the database. It
exists two different scenarios in which this could cause an issue :

- A shipment could be split through several of the companies using the
    same customs number;
- A customer is "migrating" all of their information from one of the
    companies in the database to another.

In both cases, the customs numbers can only be used in first company.

Now, the customs number is unique by company in the database.

opw-2390654

closes odoo/enterprise#15237

X-original-commit: 58fae1751fad7761ac2566a8e6fa219e5a0b7476
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>