[FIX] helpdesk: Fix days/hours conversion on SLA policies.
Currently, When we set the hours in sla policy it will be converted to
days based on standard 24 hours, not based on working calendar of the
helpdesk team.
For Ex: 32 hours is converted into 1 day and 8 hours but if working
hour is 8 then is should be converted to 4 days.

So in this commit convert the hours to days based on working hour
of the resource calendar of the helpdesk team.

closes odoo/enterprise#17289

Taskid: 2413881
X-original-commit: 6ac2ca51a809b3b7b0753fc01db11c37a48ff93d
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>