[FIX] document_spreadsheet: properly obtain spreadsheet list
The method used to retrieve the possible spreadsheets for inserting
a pivot table was not working for non-admin users.

Steps to reproduce the error: create some spreadsheets with an admin user
in a workspace that is not accessible to normal users, then with a
normal user try to use the "INSERT IN SPREADSHEET" button on a pivot
view -> Access Error.

Documents/Spreadsheet have specific access rules, and the filters are
(expected to be) consistently implemented in CRUD methods. Using direct
SQL to retrieve all spreadsheet "candidates" then manually calling
`check_access()` on the list is bound to cause access errors for all
those that are unavailable to the current user.

Fix it by using the proper CRUD methods to obtain the list of available
spreadsheets for the user. The code is also simpler.

closes odoo/enterprise#15255

X-original-commit: 6a6d363d3cf4f87f8d257afafd9e06ed213df1f8
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>