[FIX] document_spreadsheet: improve contribution detection
Spreadsheet contributors are automatically added when a spreadsheet
document is modified by a user. However, merely modifying metadata (e.g.
moving the document in a different workspace) should not be considered a
contribution.

Restrict the "contributor" detection to cases where the `raw` field is
being written (i.e. the spreadsheet is saved)

closes odoo/enterprise#15260

X-original-commit: 0ce8a218fd4e3d3f57735648003245f044325b46
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>