[FIX] approvals, documents: fix the kanban card when chatter open
Currently, Favourite and activity button display outside the kanban record
when document chatter is opened in sidebar.

So in this commit, wrap the content when kanban card is too small.
apply the class fle-wrap to display the contnet in new line for kanban bottom.

The opportunity is taken to make a pass on all of the "reload" to add the
`keepChanges` flag to prevent from losing pending changes on reload, especially
useful in form view.

closes odoo/enterprise#15345

Taskid: 2362964
X-original-commit: 429fafb42d43a290e84a57982e1bfae74b58e5a2
Related: odoo/odoo#63464
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>