[IMP] web_studio: m2m relation in custom stat button
Currently,
when creating a custom stat button through studio,
the user can only set it up by selecting a many2one field.

in this commit,
when creating a custom stat button through studio,
is able to select many2one as well as  many2many .

The older compute_function has been reworked as it used a read_group,
and it is not possible to group by many2many field. We now use
search_count instead.

closes odoo/enterprise#15461

Taskid: 2349284
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>