[FIX] social_linkedin: adapt HTTP requests to not exceed the new query size limit
Purpose
=======
LinkedIn just changed to size limit of the HTTP request (4KB max for the query parameters)
we can make to their API. We need to update the code which refresh the statistics of the
"social.live.post".

As we put the URN of the LinkedIn post and not the ID, we need ~40 bytes per "social.live.post".

Task 2418242

X-original-commit: 85763c2467fa84e5502d01ab4c67b34ec11bae2d