[FIX] planning: keep planning slot after running contract
When changing an employee's contract status to 'Running', if he already
has some planning slots based on shift templates, the end date whill be
replaced.

To reproduce the error:
(Need hr_payroll)
1. Go to Planning
2. Click on Add
3. Set the new planning slot PS:
	- Set employee E
	- Set shift template ST
	- For date, select a range of at least 2 days and keep it in
mind
4. Publish
5. Go to Payroll > Employees > Contracts
6. Select E's contract
7. Change the status to "new"
8. Change the status to "Running"
(Steps 7-8 are used to force the status "Running". This change can
appear when creating a new contract for an employee who is already
working for the company, for instance)

=> 2 undesirable behaviours can occur:
- An error message appears "Shift end date should be greather than its
start date"
- Back to Planning, find PS, the end date is different from the one
defined on step 3. The new one is only based on the start date and the
ST's duration.

Because of several `compute` variables, when changing the status of an
employee's contract, it calls the method modified in this commit.
The latter then recalculates the end date of the employee's slots
considering only the duration of the shift template and actually
ignores the defined range of days previously defined.

OPW-2411760

closes odoo/enterprise#15394

X-original-commit: 87bd8a5496b4211bf529070cbdbbf0767deba3a7
Signed-off-by: adwid <adwid@users.noreply.github.com>