[FIX] sale_amazon: Avoid dismiss pickings when amazon rate_limit_reached
Before this commit when reached the rate_limit_reached, the rest of
pickings pending to be submit to amazon submit feed were marked with
amazon_sync_pending = False making them to be ignored on the next cron
execution.

Now only the pickings already submit are market with
amazon_sync_pending=False.

closes odoo/enterprise#15489

X-original-commit: 74f60eb285d779871a5eaf45d67fa08d88f3c5ed
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>