[FIX] documents: handle space keydown on focused checkbox in list
before this commit: using keyboard navigation if we move to record from searchview
and reach to second record then it will have focus on checkbox not pressing space
on checkbox will not select record, pressing space again will select first record.

this is because pressing space key calls click handler of DocumentsListRenderer as
well as ListRenderer due to which DocumentsListRenderer's click handler will enables
checkbox and ListRenderer's click handler will uncheck it.

after this commit: as we handle click handler specifically in DocumentsListRenderer
so we have to handle space key and stopPropagation else it will call ListRenderer's
click handler as well.

task-2411686

X-original-commit: bedaeb0dc89b56d560076cddfc6fd5631e0d0cb4