[IMP] industry_fsm_sale: add searchpanel on fsm task's product view
Purpose of the task is, It is not easy to find the right product when
they are hundreds of them so a search panel on product category would
help the user quickly access the products.

So in this commit, Added a search panel on product category when going
to add the product on fsm task.

TaskID: 2417252
Closes: #15550