[FIX] social_push_notifications: fix issue if no visitor timezone
When a social.post is sent using the "use_visitor_timezone" feature, we compare
the social.post's creating user timezone with the visitors' timezones.

However, there can be cases where the visitor was not assigned a timezone.
Before this commit, we used "self.env.user.tz" as a fallback, but the posts can
be potentially sent by a CRON where the environment user has no timezone
either, leading to a crash.

To avoid that failure, we now fallback to the neutral 'UTC' timezone, which is
as good as any when we have no clue what the actual visitor's timezone is.

A test was added to ensure that a visitor without a timezone doesn't create a
raise.

Task-2415790

closes odoo/enterprise#15684

X-original-commit: f1a4ede8decdce33bdc257378762444df60d223c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>