[FIX] quality_mrp: forbid to set as done WO with QC
When a user uses the simplified view to process a Work Order which has
quality checks to be done, there is currently no warning about it when
marking the WO as done using the "Done" button of the simplified view.
The user will therefore get an error when trying to "Mark As Done" a
Manufacturing Order even if all the Work Orders are set as "Done"
without warnings.

This commit forbids to set as "Done" in the simplified view a workorder
for which a Quality Check should be done, sending a warning message to
complete the QC. This does not apply for Quality Checks of type
Instructions, Register Component and Register by Product. For those
types, it is possible to set the WO as Done and this will also set the QC as
Done.

closes odoo/enterprise#19631

Signed-off-by: Arnold Moyaux <amoyaux@users.noreply.github.com>