[FIX] web_studio: fixed `show invisible elements` checkbox issue
Currently, 'showInvisible' getter function is called only when user reload the page
in studio mode due to that 'this.showInvisible' always return the false value
becuase of that when user switchback to view tab after checking 'Show Invisible Elements'
checkbox, does not show the effect on the view.

So in this commit convert getter function to regular function to recompute the
'show_invisible' value each time.

TaskID:2358101

closes odoo/enterprise#16352

X-original-commit: af9a3d0a7a0cee13689237cfd5623bfcccaf40ba
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>