[FIX] hr_appraisal: "View Appraisal" mail button
When sending an appraisal request, if the sender uses a template that
contains the "View Appraisal" button, the latter will not be displayed
in the mail.

To reproduce the error:
1. Go to Settings > Appraisal
2. In "Confirmation Email Template", open the one for the employees
	- You can see that the template contains a "View Apparaisal"
button
3. Go to Employees and select Marc Demo
4. Request Appraisal

=> The button is not shown

One field of the rendering context was badly defined: `user_id` is "The
internal user in charge of this contact". Here, we need `user_ids`, the
one2many field that links a `res.partner` with the corresponding
`res.users`.

OPW-2418244

closes odoo/enterprise#15490

X-original-commit: 4307a1b96a263e4c5bb059b10c041a05ee2b68f5
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>