[FIX] voip: avoid permanent override of the back button
The commit odoo/enterprise@b9c50fd28f162f4cba01c07c05bdc5bced1aeb78
enabled the back-button support for VoIP's DialingPanel.

Sadly when the `voip` module is installed, the DialingPanel is
immediately instantiated by the WebClient and appended to the DOM at
launch.

Once attached to the DOM, the `BackButtonEventMixin` overrides the
default back-button behavior... but in this case, this override should
happen only when the DialingPanel is displayed, not just attached to
the DOM.

In practice, when the `voip` module is installed, the back-button
behavior is overridden all the time by the DialingPanel widget.

This commit fixes it by only enabling/disabling the back-button override
when the widget is actually displayed (independently of its presence in
the DOM).

Task ID 2148384

closes odoo/enterprise#15594

X-original-commit: 19903365dc0562b95570aa4e8ed5fe8e607b166c
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>