[FIX] industry_fsm: Activity does not show in listview.
Before this commit:

when try to see activity in fsm list view, it does not show any activity.
but for this record, if you check in kanban view, it shows activity.

Activity does not display due to wrong field for widget list_activity
so set the correct field as 'activity_ids' to show the activity

closes odoo/enterprise#15732

Taskid: 2422724
X-original-commit: e22205e329cbfe72efe3ba5f21f61eef5d8fb18a
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>