[FIX] account_reports: set active_id when opening journal items
Behavior prior to this commit:

When you open the Journal Items from a financial report (for example P&L
report), then view one of the entries and go back (or reload the window
with the journal items), a SQL error is shown:

    psycopg2.DataError: invalid input syntax for integer: "financial_report_group_4_21"
    LINE 1: ...company_id" in (1))) AND "account_account".id IN ('financial...

Behavior after this commit:

Able to go back to the journal items, or refresh the journal items
window, without a SQL error

Note:

The error was caused by the fact that the `active_id` was defaulted to
the current view id, i.e. the report XML id, a string, but the action
(`action_move_line_select`) has a context of:
`{'search_default_account_id': [active_id]}`.  The system was thus
attempting to use the string "financial_report_group_4_21" as an account
id.  To work around the account id, as passed to the open_journal_items
method, was returned as `active_id` key in the result context.

opw-2416831

closes odoo/enterprise#15499

X-original-commit: 050455f8521477c155bdfb0e819d421bb5c481f7
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Nicolas Galler <ngaller@users.noreply.github.com>