[IMP] helpdesk_sale_timesheet: update get last SOL of customer
Before this commit, we search and retrieve a list of SOL and then we
filter this list to find the last SOL with the remaining hours is
greater than 0.

Since, the remaining_hours field is now stored in sale.order.line
model, we can update this feature to do it only in the search and find
only the SOL that we expect.

task-2424382