[IMP] helpdesk_timesheet: compute sudo and store total_hours_spent
Before this commit, the total hours spent is only the number of hours
spent in the timesheets that the user can see and not all timesheets
contained in the ticket.

This commit adds compute_sudo in this field and stores this field in db
to get the correct total hours spent that we expected. That is, the
total hours spent of all timesheets in the ticket.

task-2424382