[FIX] account_asset: change deferred expenses value
On a deferred expense, when editing the expense of a line, if at least
one has been posted, the "Next Period Expense" field of the next lines
won't be correctly computed.

To reproduce the error:
(Hint: use demo data)
1. Go to Accounting > Accounting > Management > Deferred Expenses
2. Create a new one, e.g.:
	- Original value: $10,000
	- Acquisition Date: 01/01/2020
	- Number of recognitions: 5 Months
	- Deferred Expense Journal: 141000 Prepayments
	- Expense Account: 961000 RD Expenses
	- Journal: Miscellaneous Operations (USD)
3. Save, Confirm
4. Select the first depreciation line, Post it
5. Refresh the page
6. Click on Edit
7. Select the second line to update it:
	- Set the Expense field to $1,000 (instead of $2,000)

=> The "Next Period Expense" field of the selected line is now $5,000
but it should be $7,000

The system based the computations on the `value_residual` field of the
deferred expense. Since one line has been posted, this field is worth
$8,000. It then removes the first entry ($2,000) and the new value for
the second line ($1,000). This is the reason why it displays $5,000.
Obviously, due to a cascade effect, the next lines will also be incorrect.

OPW-2409086

closes odoo/enterprise#15533

X-original-commit: 4ab1d32a6f3cc29cf02fb992e6536c2ec37e855c
Signed-off-by: Adrien Widart <adwid@users.noreply.github.com>