[FIX] helpdesk_sale_timesheet: condition always False
Before this commit, we check if the bill_type in the project is equal
to 'customer_type' but this field can only be either 'customer_task' or
'customer_project'. Thus, this condition will always be False.

This commit changes the condition to check if the bill_type is equal
to 'customer_project'. Then, this condition can be True if the project
has a SOL and his bill_type field is equal to 'customer_project'.

closes odoo/enterprise#15540

X-original-commit: a493b6f00260a40f6f9b69fef06cd0fe06ad8a03
Related: odoo/odoo#64012
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>