[FIX] account_asset: fix asset on 31st december
The account assets tests failed on 31st december in part because we fake
today because of special case of 31st december but we don't fake
context_today (used since b13f40c83e for auto_post for example).

Also adding some override of today that prevented test to pass on 31st
december.

closes odoo/enterprise#15517

X-original-commit: ec9f3cc6c31036a0c5eb7a8a0e44718e7ed9a538
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>