[FIX] web: scrollbar when html/o2m is first child of page
before this commit: when html or o2m widget is first child of first notebook
page of form which does not have sheet tag, it shows unnecessary horizontal
scrollbar because of negative margin of 32px given in scss mixin which should
be applied only when there is a sheet tag in form but negative margin of 32px
is also applied when there is no sheet tag and due to which negative margin is
applied and horizontal scrollbar comes.

after this commit: when there is no sheet tag and html or o2m widge is first
child of first notebook page then negative margin of 32px will not applied,
negative margin of 32px will be applied only when there is a sheet tag.

task-2411627

closes odoo/enterprise#15519

Related: odoo/odoo#63961
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>