[FIX] hr_work_entry_contract: properly initialize contract_id
Strap down, this one is a mess:
    - Install hr_payroll with demo data
    - Uninstall hr_work_entry_contract (on which hr_payroll depends)
    - Reinstall hr_work_entry_contract
    -> The ORM will complain that the field 'contract_id' cannot be set
    to NOT NULL because NULL values exist.

Here's a quick overview of the way in which these modules depend on each
other:

    hr_work_entry <-- hr_work_entry_contract <-- hr_payroll

hr_work_entry_contract will introduce the `contract_id` field on model
`hr.work.entry` which is required.

`hr_payroll` calls `generate_work_entries` in order to create demo
records of `hr.work.entry`, these `hr.work.entry` records use the
extended version of the model introduced by `hr_work_entry_contract`,
meaning the required `contract_id` is provided.

Since the demo records were created using a generic function on
`hr.employee`, no corresponding `ir.model.data` fields are generated,
thus when `hr_work_entry_contract` is uninstalled (i.e. `hr_payroll` is
uninstalled), the demo records are not unlinked, they linger in the
database and the `contract_id` column is dropped.

Upon reinstalling hr_work_entry_contract, the ORM will fail to set the
field `contract_id` to `NOT NULL` because of the existing demo data from
`hr_payroll`.

The best fix I could find for a stable version was to simulate the
behaviour of `generate_work_entries`, except instead of generating new
work entries it simply tries to match existing ones and update their
contract_id from NULL to a valid value during the column's
initialisation.

In master though, better fixes include, but are not limited to:
    - Not using a function to create demo data and instead creating
    records by hand in the XML file.
    - If using a function to create demo data, making sure to create the
    corresponding ir.model.data so that the ORM can do automatic cleanup
    upon uninstall.
    - Making `contract_id` on `hr.work.entry` not a required field.

Note that only the last of the three proposals would fix the scenario in
which a real customer in production has multiple existing
hr.work.entry and decides to install hr_work_entry_contract or
hr_payroll, and in fact this fix does not completely handle that
scenario either, it's a best effort that might work if a contract exists
for a given employee and it happens to be valid during the existing work
entries.

closes odoo/enterprise#15537

X-original-commit: 5a7aa743b1ba068def9de1356807057ca3097a4c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Adrian Torres (adt) <adt@odoo.com>