[FIX] account_accountant: Fix writeoff in manual widget in multi-currency
When reconciling using the manual reconciliation widget using a foreign currency and adding a writeoff, the generated writeoff journal entry was wrong because the journal items got an invalid amount_currency sign regarding the balance.
Indeed, if balance is >= 0 (resp. <=), amount_currency must be >= 0 (resp. <=).

Before this commit, having an invoice and a payment with these journal items:
|              | Account       | Amount Curr   | Balance       |
| ------------ | ------------- | ------------- | ------------- |
| Invoice      | Receivable    | 1000 EUR      | 500 USD       |
| Payment      | Receivable    | -800 EUR      | - 400 USD     |

Was generating these write-off lines in the manual reconciliation widget:
| Account       | Amount Curr   | Balance       |
| ------------- | ------------- | ------------- |
| Receivable    | 200 EUR       | 100 USD       | <- should be -200
| Write-off     | -200 EUR      | 100 USD       | <- should be 200

Of course, such journal items trigger the sql constraint on account.move.line leading to a UserError returned to the user.

closes odoo/enterprise#15591

Task: 2425638
X-original-commit: 8cae761a5735ca7a874dca35c500f7029241f651
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>