[FIX] account_followup: allow to partially process records
Have a first followup level to send mail in auto execute (the default
'First reminder email' will do)
Have a list of partners in need of action in the followup menu
One of the partner need to have the email field empty (no email)
Let the scheduled action 'Account Report Followup; Execute followup'
process the followups via scheduler or run manually.

No mail will be sent even if all other partners data is correct.
This occur because the scheduler action is run in a method which will
raise on any error, aborting the execution.

opw-2379650

closes odoo/enterprise#15596

X-original-commit: 6754c4cc78adc2405bf250009db8bc5e3f6926e8
Signed-off-by: agr-odoo <agr-odoo@users.noreply.github.com>