[IMP] web_mobile: BackButton hook for OWL
This commit introduces a hook to allow OWL component to implement a
custom handling of the "backbutton" event, available in the mobile app.
It achieves the same purpose as the existing `BackButtonEventMixin`.

To do so, it required a refactoring of the `BackButtonEventMixin` to
extract the main handling of the backbutton events to a common place,
usable by both implementation (Widget & OWL), in the name of the
`backButtonManager`.

This hook will be used later on in modules like 'mail', 'web'...

Note: if multiple listeners are registered, they will be called
following LIFO (Last In, First Out) methodology.

Related to task ID: 2347640

X-original-commit: 96beff3d22e7083ec7a51d29fc6fdaef29cee912
Co-authored-by: Sébastien Theys <seb@odoo.com>