[IMP] helpdesk_sale_timesheet: restrict helpdesk projects to task rate
Before this commit, the pricing of a project linked to at least a team
can be employee, project and task rate. But in reality, the task rate is
most adapted, because the customer is often different in each ticket of
a helpdesk team. For this reason, in this commit, we restrict the
pricing type of projects linked to at least a helpdesk team to task rate
only.

task-2424382