[FIX] industry_fsm_sale: remove the SO if it is a fsm project
Before this commit when we update a project to convert into a fsm
project, the SO defined in this project is not removed. Since we
restrict the pricing type to task rate when it is a fsm project, the SO
in the project is useless.

This commit removes the SO if is_fsm field is set to True.

task-2424382