[FIX] helpdesk: fix overview display when grouping by in helpdesk team
Currently, when helpdesk dashboard grouped then it is not displayed
below the dashboard value, it is displayed besides it so grouped kanban
view is not displayed and screen has horizontal scroller, to view
kanban view user have to scroll screen

So in this commit, Append the dashboard before the kanban view to display
it outside the kanban so it does not break the grouped kanban view.
also hide the 'Add new column' on group so it should work properly.

closes odoo/enterprise#16804

Taskid: 2424177
X-original-commit: 9a1ae807835d5b79c5d64d3224a2fed1b323202b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: IT-Ideas <ltu@odoo.com>