[FIX] web_mobile: crash when opening a Dialog in debug mode
When the mobile QUnit test suite is ran in debug mode, the test "dialog
is closable with backbutton event" throws a "This listener was already
registered." error.

Since commit odoo/enterprise@96beff3d22e7083ec7a51d29fc6fdaef29cee912
the BackButtonEventMixin (via its BackButtonManager) prevents from
registering multiple times the same Widget/Components's instance as a
listener for the "backbutton" event (available in the mobile app).

But it raises a hidden error present in the original implementation of
the "backbutton" event support for the Dialog widget : the
`on_attach_callback` method can be called multiple times due to its call
being made in the `opened()` function instead of the internal `_opened`
property.

This commit fixes it by ensuring that the `on_attach_callback` is called
only when the `_opened` promise is resolved and not every time we try to
access it via the `opened` function.

closes odoo/enterprise#15631

X-original-commit: e4e932bf3b1a88590566b06c70e8a39f6b5fbe97
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>