[IMP] web_enterprise: add simple tap to add an event (day/week)
Before this commit, the only way to add an event was using a long tap.

After this commit, a short tap will be added.

So to summarize on Mobile there is two methods to add an event using the
calendar: short tap and long tap.

For month/year views when a user tap on a "Free Zone" (no event)
the view is changed in the "day view" for this clicked day.

Task ID: 1891957