[FIX] sign: make customer field usable on mobile
In desktop, some tweaks has been done to an editable list view
to transform a record value in a label....
See .o_sign_flat_o2m.

This customization allow us to directly fill in the partner_id field
via signer_ids and having the impression that it belongs to the form
view.

But it didn't worked on mobile.
To be clear, this code will have to be deleted in the future.
The good solution would be to find a way to have 'partner_id' field
directly in 'sign.send.request'.

closes odoo/enterprise#15662

Task-id: 2387621
X-original-commit: 776676cc528211def291737a9f83f7d6bd5def87
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>