[FIX] helpdesk_sale_timesheet: processing when update sol in timesheets
Before this commit, when the user changes the SOL in a timesheet in
ticket view form for instance, then we retrieve this update to recompute
the qty_delivered in old sol. The problem is we do not check if the
value in timesheet is a list and if it is tuple/list in this list. If it
is not the case, we have a traceback.

This commit checks if the value of timesheet_ids in vals is list/tuple.
Then we also check if the containing in this list is only list/tuple
too. Finally, we can do the processing to search all old sols to
recompute the qty_delivered.

closes odoo/enterprise#15660

X-original-commit: e28d167663242a604f8365b32e6f224a1662ce53
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>