[FIX] account_batch_payment: Fix 'amount' field of batch payments
The amount was computed as the sum of payment 'amount'.
However, 'amount' in payment is always positive even if the payment is outbound.
Even the batch is rectricted to be full inbound/outbound, there are some flows creating batch payment mixin both of them.
For example, it happens when refunding credit card transactions. Suppose you made twice the same payment using Ogone, then you refund one of them. At the end of the month, Ogone will give you an excel document representing the batch you will create and the transaction details including the inbound and the outbound ones.
In that case, the batch 'amount' is wrongly computed.

closes odoo/enterprise#15675

X-original-commit: 0172e913648b2767980e4b464763593db89ab145
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>