[IMP] account_invoice_extract: use webhooks to update the requests
Previously, we were only using a cron that ran every 30 minutes to
automatically update the extraction requests.

The client will now send a webhook URL as part of the request that
the extraction server will call once the results are ready.

The cron still remain with a much lower frequency. It acts as a
backup in case something went wrong during the webhook call.

Task: 2416741

Related server-side task: 2452019

closes odoo/enterprise#15663

Signed-off-by: Florian Daloze (fda) <fda@odoo.com>