[IMP] helpdesk_timesheet: use access rights if user is internal
Before this commit, when the user can only see his own timesheets, this
user can see more than his own timesheets in timesheet portal
(/my/timesheets). It is because we only use the domain defined for
portal and we do not use the one defined in acces rights for internal
users.

This commit uses the domain defined in access rules for the internal
users. If it is a portal user than we use the domain for portal as
before.

task-2424382