[IMP] industry_fsm_sale: move sql constraint for timesheet_product_id
Before this commit, we want the timesheet_product_id field be set if the
project is billable and allows timesheets.
But, now the timesheet_product_timesheet field is only required for fsm
project which is billable and allows timesheets.

This commit adds the sql constraint created in sale_timesheet in
industry_fsm_sale module and update the constraint for this field will
be required for fsm project which is billable and allows timesheets.

task-2424382