[ADD] helpdesk_fsm_sale: link for helpdesk_sale_timesheet and fsm sale
Before this commit, we cannot hide 'Invoice' page if
helpdesk_sale_timmesheet and industry_fsm_sale modules are installed.
Because in both module, we want to restrict the pricing type to task
rate (for helpdesk projects and fsm projects). But the
timesheet_product_id must be visible and required and then do not hide
the 'Invoice' page for fsm projects, compare to helpdesk projects,
we do not want to display this field and then hide 'Invoice' page.

This module is use to do the bridge between the industry_fsm_sale and
helpdesk_sale_timesheet modules to hide the 'Invoice' page for helpdesk
projects and show this page and only timesheet_product_id field for fsm
projects in project view form.

task-2424382