[FIX] sale_renting: use ir.lang datetime format in line descriptions
Steps:
- Enable debug mode
- Go to Settings > Translations > Languages
- Edit "English (US)" and change the Date Format to "%d/%m/%Y"
- Go to Rental
- Create a new Rental Order
- Add any product that can be rented
- Select a period
- Click Add

Bug:
The dates in the description of the line are not displayed using the
format set earlier.

Explanation:
Passing `dt_format='short'` to `format_datetime()` bypasses date format
customization. The only way to use the customization is to pass it
`dt_format=False` as seen here:
https://github.com/odoo/odoo/blob/bda051d1a681d149f197675516ade1c731391c5a/odoo/tools/misc.py#L1300-L1303
The same applies to `format_date()` and `format_time()`.

This commit fixes date and time formatting for new Rental Orders.

opw:2429366

closes odoo/enterprise#15727

X-original-commit: 9966b2a4eb7f4ed87ffaf39b65f30b406996c7d5
Signed-off-by: backspac <backspac@users.noreply.github.com>