[FIX] l10n_be_hr_contract_salary: fix warning condition
An error should be raised when someone sign the salary configurator
if the car is not available anymore.

Before this commit, the check for available car based on the driver
was wrong.

Use the _get_available_vehicles_domain to check if the car taken
is still available when the salary configurator is signed.

closes odoo/enterprise#15719

X-original-commit: 6bfd5251a963cdb5d112145ca70f38bd3abf24ba
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>