[FIX] account_online_synchronization: fix some problems
- When searching for the last added transactions in the system,
the search was not correct as it was searching on all entries.
It should search on entries on a particular journal and with the same
account.online.account. This caused problem in case there was multiple
synchronization from different banks and one of them uses a provider
relying on that information (ponto in this case). Also the default
sort of account.bank.statement.line is by date asc which in our case
is wrong as we want the latest entry before a given date.
- Balance was never updated because it was written after the break from
the loop.
- Also add db_uuid to all call so that we can prevent error in case
database has been duplicated by the user on saas/paas (db_uuid is
different when duplicating). The error happening when duplicating
database is the following: when one database renew its refresh_token
the second database has a signature error and can't connect anymore.
As a matter of fact, online synchronization is not supposed to work
with the same credentials on two different database.

closes odoo/enterprise#15796

X-original-commit: 3e245cf300bbb8a92916b0be46c9636221b7fd09
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>