[IMP] helpdesk_sale_timesheet: replace _get_not_billed to _is_not_billed
The _get_not_billed method used to filter timesheets and to have only the
ones which are not billed has been replaced by _is_not_billed in
sale_timesheet module. Because, we always use this method after a
filtered method. To avoid to have successive filtered, we change this
method to process only on one recordset in self and return True if the
timesheet is not billed, otherwise return False. With this changes, we
can use the method in the first filtered.

task-2424382