[IMP] industry_fsm_sale: use related timesheet_product_id in task
Before this commit, we do some checks with timesheet_product_id of
project in project.task. But we have timesheet_product_id field in
project.task which is now a related field of the one in project.

This commit replaces the occurence of project_id.timesheet_product_id
in task to use the related field in this model.

task-2424382