[IMP] helpdesk_timesheet: change project and remove task fields in ticket
Before this commit, in a ticket, we could change a project and add a
task to a ticket. But the mecanism is not very useful, it's better to
use the integration with industry_fsm app to create tasks from a ticket.

This commit replaces the project stored field in ticket to a related
field linked with the project_id field in helpdesk.team. Then, it
removes the task_id field in helpdesk.ticket and removes all useless
methods linked to this removed field.

Moreover, we need to store this field because there is a read_group in
project.project to retrieve the number of ticket in a project.

task-2424382