[FIX] sale_amazon: require to ship all components of a product at once
Before this commit, having a product configured as both an Amazon
product (SKU set as the internal reference) and a kit linked to a BoM
would eventually cause delivery confirmation issues.

Take for instance a kit product linked to a BoM consisted of multiple
components. If only 1 component is marked as 'done' when the delivery
is confirmed, the Connector will notify Amazon that the related
product has been shipped. Subsequent confirmations of other components
of that product will also send a notification.
This is problematic because Amazon sends a confirmation email to the
customer who will in fact not receive the entire product components at
once. Furthermore, Amazon might discard the other notifications.

As of this commit, a `UserError` is raised if not all components of a
given product are marked as 'done' at once when the delivery is
confirmed. If a delivery contains move lines related to different
products, it remains possible to only confirm moves related to some
products without having to mark all components of all products as
'done'.

opw-2411279

X-original-commit: 749c5d9bca20be269ac6cbc436c5f4a4bbe5f547