[FIX] quality_control: ignore check of absent products
- Install stock, quality_control and purchase
- Go to Inventory > Configuration > Settings and enable:
  * Lots & Serial Numbers
  * Storage Locations
  * Multi-Step Routes
- Go to Inventory > Configuration > Warehouses and configure Warehouse
by setting "Receive goods in input, then quality and then stock (3 steps)"
as "Incoming Shipments"
- Create 2 storable Products tracked by Lots (i.e. Product X and Product Y)
- Go to Quality > Quality Control > Control Points and create Quality Control
Points for Product X and Product Y with:
  * Operation: Internal Transfers
  * Type: Measure
- Go to Purchase and create a PO with Product X and Product Y
- Confirm order and validate reception for Product X only
- Create a Backorder for Product Y
- Go to Inventory and open Internal Transfers page
- Open Transfer to "WH/Quality Control" created from PO
- Launch Quality Checks
The measure check will be asked for Product X, but also for Product Y,
which will not be possible to perform as it has not been received yet.

The Quality Checks are created when PO is confirmed, based on ordered
Products.
As some Products can be missing at reception, the Quality Check for these
Products should be ignored.

opw-2409735

closes odoo/enterprise#15828

X-original-commit: 1482fc449012b29e4136c22cd8519a5e90d821b3
Signed-off-by: Anh Thao PHAM <kitan191@users.noreply.github.com>