[FIX] l10n_mx_reports: 251.04.xx accounts not recognize in MX COA report
If an account like "251.04.01" with the tag "251.04 Acreedores diversos a largo plazo nacional parte relacionada"
is present in a mexican COA report, the account will be considered as misconfigured and the export for SAT
button will trigger an error.

In the code, the domain used in the report for the code "MX_COA_251" contained two times the tag "251.03" and
no mention of "251.04". Thus the tag "251.04" was added in the report and now accounts with this tags are no
longer considered as misconfigured.

opw-2424711

closes odoo/enterprise#15846

X-original-commit: 506e96afe7fe214e99461dc146a8ba1a2ca6d7e0
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>