[IMP] sign: improves cc feature on sign
Improves and fixes the CC feature of sign module. Changes the view of the sign wizard to have two notebook pages containing a message field for signers and one message field for cc followers. In sign now wizard, a message field for cc is added.
Also improves the mail template for signer and cc. Removes the email template for task follow.  All changes were made with the goal of leaving the sign flow more clear for the user.

task-2375039.

closes odoo/enterprise#15812

Related: odoo/upgrade#2217
Signed-off-by: Arnaud Joset <arj-odoo@users.noreply.github.com>